support for floating-point numbers added(McMMO/Jobs)

+++ src/main/java/com/tamrielnetwork/gbooster/gbooster/bar/BoosterBar.java
@@ -45,10 +45,10 @@ public class BoosterBar {
 
     private String getTitle(){
         return Utils.replaceColors(Objects.requireNonNull(main.getConfig().getString("bar-pattern"))
-                .replace("%minecraft%", String.valueOf(main.getActiveBoostersManager().getBoosterMultiplier(BoosterType.MINECRAFT, false)))
-                .replace("%mcmmo%", String.valueOf(main.getActiveBoostersManager().getBoosterMultiplier(BoosterType.MCMMO, false)))
-                .replace("%jobs_xp%", String.valueOf(main.getActiveBoostersManager().getBoosterMultiplier(BoosterType.JOBS_XP, true)))
-                .replace("%jobs_money%", String.valueOf(main.getActiveBoostersManager().getBoosterMultiplier(BoosterType.JOBS_MONEY, true)))
+                .replace("%minecraft%", String.valueOf((int)(main.getActiveBoostersManager().getBoosterMultiplier(BoosterType.MINECRAFT, false))))
+                .replace("%mcmmo%", String.valueOf((int)(main.getActiveBoostersManager().getBoosterMultiplier(BoosterType.MCMMO, false))))
+                .replace("%jobs_xp%", String.valueOf((int)(main.getActiveBoostersManager().getBoosterMultiplier(BoosterType.JOBS_XP, true))))
+                .replace("%jobs_money%", String.valueOf((int)(main.getActiveBoostersManager().getBoosterMultiplier(BoosterType.JOBS_MONEY, true))))
                 .replace("%duration%", String.valueOf(main.getActiveBoostersManager().getMostOldBoosterInMinutes())));
     }
